requirements.txt
Django==5.2
gunicorn==23.0.0
whitenoise==6.9.0
python-dotenv==1.1.0
authlib==1.6.0
PyJWT==2.10.1
langchain-google-genai==2.1.4
langchain-core==0.3.60
Pillow==10.4.0
requests==2.32.3
dj-database-url==2.3.0
psycopg2-binary==2.9.10